refactor(frontend): extract component to add entity to collection
