+++ apps/frontend/src/pages/media/item/index.tsx
@@ -1,4 +1,5 @@
 import {
+	AddEntityToCollectionModal,
 	MediaScrollArea,
 	PartialMetadataDisplay,
 	ReviewItemDisplay,
@@ -31,7 +32,6 @@ import {
 	NumberInput,
 	Paper,
 	ScrollArea,
-	Select,
 	SimpleGrid,
 	Slider,
 	Stack,
@@ -45,8 +45,6 @@ import { DateInput } from "@mantine/dates";
 import { useDisclosure, useLocalStorage } from "@mantine/hooks";
 import { notifications } from "@mantine/notifications";
 import {
-	AddEntityToCollectionDocument,
-	type AddEntityToCollectionMutationVariables,
 	CreateMediaReminderDocument,
 	type CreateMediaReminderMutationVariables,
 	DeleteMediaReminderDocument,
@@ -70,7 +68,6 @@ import {
 	SeenState,
 	ToggleMediaMonitorDocument,
 	type ToggleMediaMonitorMutationVariables,
-	UserCollectionsListDocument,
 	UserMediaDetailsDocument,
 	UserReviewScale,
 } from "@ryot/generated/graphql/backend/graphql";
@@ -239,81 +236,6 @@ const MetadataCreator = (props: { name: string; image?: string | null }) => {
 			</Text>
 		</>
 	);
-};
-
-const SelectCollectionModal = (props: {
-	opened: boolean;
-	onClose: () => void;
-	metadataId: number;
-	refetchUserMedia: () => void;
-}) => {
-	const [selectedCollection, setSelectedCollection] = useState<string | null>(
-		null,
-	);
-
-	const collections = useQuery({
-		queryKey: ["collections"],
-		queryFn: async () => {
-			const { userCollectionsList } = await gqlClient.request(
-				UserCollectionsListDocument,
-				{},
-			);
-			return userCollectionsList.map((c) => c.name);
-		},
-	});
-	const addMediaToCollection = useMutation({
-		mutationFn: async (variables: AddEntityToCollectionMutationVariables) => {
-			const { addEntityToCollection } = await gqlClient.request(
-				AddEntityToCollectionDocument,
-				variables,
-			);
-			return addEntityToCollection;
-		},
-		onSuccess: () => {
-			props.refetchUserMedia();
-			props.onClose();
-		},
-	});
-
-	return collections.data ? (
-		<Modal
-			opened={props.opened}
-			onClose={props.onClose}
-			withCloseButton={false}
-			centered
-		>
-			{collections ? (
-				<Stack>
-					<Title order={3}>Select collection</Title>
-					{collections.data.length > 0 ? (
-						<Select
-							data={collections.data}
-							onChange={setSelectedCollection}
-							searchable
-						/>
-					) : undefined}
-					<Button
-						data-autofocus
-						variant="outline"
-						onClick={() => {
-							addMediaToCollection.mutate({
-								input: {
-									collectionName: selectedCollection || "",
-									entityId: props.metadataId,
-									entityLot: EntityLot.Metadata,
-								},
-							});
-						}}
-					>
-						Set
-					</Button>
-					<Button variant="outline" color="red" onClick={props.onClose}>
-						Cancel
-					</Button>
-				</Stack>
-			) : undefined}
-		</Modal>
-	) : undefined;
 };
 
 const CreateReminderModal = (props: {
@@ -1255,11 +1177,12 @@ const Page: NextPageWithLayout = () => {
 										<Button variant="outline" onClick={collectionModalOpen}>
 											Add to collection
 										</Button>
-										<SelectCollectionModal
+										<AddEntityToCollectionModal
 											onClose={collectionModalClose}
 											opened={collectionModalOpened}
 											metadataId={metadataId}
 											refetchUserMedia={userMediaDetails.refetch}
+											entityLot={EntityLot.Metadata}
 										/>
 									</>
 									<Button
